Extract `DateFormatter` RFC-3339 to file and rename to `wordPressCom`
